Move all user-interaction logic to file
Includes scrolling and gesture recognizers
